Tests/LibWeb/Text/expected/HTML/DOMParser-parseFromString-srcdoc-base-url.txt

iframe URL: about:srcdoc
iframe base URI inherited: true
parsed URL: about:srcdoc
parsed base URI: about:srcdoc
